requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
coverage==7.2.7
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==4.1.9
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
executing==1.2.0
filelock==3.12.0
Flask==2.3.2
idna==3.4
inflection==0.5.1
ipykernel==6.22.0
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2023.4.29
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pygame==2.1.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2023.3
pyzmq==25.0.2
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.2
traitlets==5.9.0
urllib3==2.0.2
virtualenv==20.23.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
Werkzeug==2.3.4
